pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "resume-tailor"
version = "0.1.0"
description = "AI-powered resume tailoring tool with file upload support"
readme = "README.md"
authors = [
    {name = "Aravind Adari", email = "aravind.adari@gmail.com"},
]
requires-python = ">=3.8"
dependencies = [
    "openai>=2.6.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
    "python-docx>=1.2.0",
    "PyPDF2>=3.0.1",
    "click>=8.0.0",
    "fastapi>=0.120.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.3",
    "httpx>=0.27.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=1.0",
]
license = {text = "MIT"}

[project.scripts]
resume-tailor = "src.main:main"